fix: resolve CI failures - linting, type checking, and build errors
- Fixed 82 ruff linting errors (unused imports, import sorting)
- Fixed 41 mypy type checking errors (missing awaits, type annotations)
- Fixed pysqlcipher3 build failure (moved to optional dependencies)
- Added missing logger imports and async wrappers
- 290/295 tests passing (98%)

Remaining: 16 non-critical style warnings, 7 minor type issues

+++ src/self_improvement/metrics.py
@@ -172,7 +172,7 @@ class ImprovementMetrics:
                 "XL": 80,
             }
             total += effort_map.get(effort, 8)
-        return round(total)
+        return int(round(total))
 
     def _count_by_effort(self, findings: list[dict]) -> dict:
         """Count findings by effort level."""
